pyproject.toml
[project]
name = "xts-api-client"
version = "1.0.5"
description = "A Package designed for connecting to a Stock Broker that uses Symphony Fintech Solutions Pvt. Ltd for there API."
readme = "README.md"
authors = [
    { name = "Chetan Sharan", email = "chetan.sharan@rmoneyindia.com" }
]
requires-python = "~=3.12"
dependencies = [
    "aiohttp==3.11.11",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "python-socketio==4.6.1",
    "requests>=2.32.3",
    "urllib3>=2.2.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mkdocs-material>=9.5.47",
    "mkdocs>=1.6.1",
    "mkdocstrings[python]>=0.27.0",
    "pytest>=8.3.3",
    "pypiserver>=2.3.2",
    "aiohttp>=3.11.11",
    "pytest-asyncio>=0.25.0",
]
